pyproject.toml
[project]
name = "spectrometer"
version = "0.1.0"
description = "Bias Spectrometer — extract training bias from AI models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "alembic>=1.13",
    "stripe>=11.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "email-validator>=2.0",
    "python-dotenv>=1.0",
    "openai>=1.40",
    "requests>=2.31",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "httpx>=0.27"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
